Add model argument transforms (#6)
* remove unused initialze method

* comment tweaks

* cache model argument transformations

* make residuals options for linear risk factor models

* consolidate Cox model constructor

* casefold only first character of prop name

Person property names like 'totChol' exist. I don't think this is safe
for all codepoints supported by Unicode, but I hope to improve upon it
in future iterations.

* add param to transforms dict even if empty

somtimes property names differ from param name (e.g., lagSbp). (This
appears to have been a bug from the original code, but will leave it in
for now and as Jim.)

* remove unused function

* sort categorical params in model constructor

* add categorical transforms

* use categorical tranforms in model eval

* remove unused function

* append 'recursive' to 'get_argument_transforms'

* add iterative impl of get_argument_transforms

* use iterative impl. of get_argument_transforms

tests seem to be passing! Not sure if it's cleaner, but it's probably
the better than the recursive implementation for broader use.

* add transform callables

* formatting

* more formatting

* move model arg transforms to own file

* remove recursive impl.

* move get_argument_transforms to transforms file